Make route url ordering predictable through alphabetical sorting
When placeholders use route indices these need to be predictable.
Otherwise the value of the placeholder will change from time to time.

We opt for an alphabetical order for now. At least this should leave us
in a situation where only changes to the application routes could lead
to route index changes.

Update tests accordingly.